third_party/blink/web_tests/platform/win/external/wpt/dom/events/non-cancelable-when-passive/non-passive-mousewheel-event-listener-on-body-expected.txt

This is a testharness.js-based test.
FAIL non-passive mousewheel event listener on body assert_equals: expected true but got false
Harness: the test ran to completion.
